completed slice that fetches coin data and updates state. layout has a provider that wraps every component/page
