+++ src/app/utils/apiData.ts
@@ -1,8 +1,17 @@
 import axios from "axios";
 
 export const fetchCoins = async () => {
-  const response = await axios(
-    `https://api.coingecko.com/api/v3/coins/markets`
+  const response = await axios.get(
+    `https://api.coingecko.com/api/v3/coins/markets`,
+    {
+      params: {
+        vs_currency: "usd",
+        order: "market_cap_desc",
+        per_page: 20,
+        page: 1,
+        sparkline: false,
+      },
+    }
   );
 
   return response.data;
