+++ src/components/Table.tsx
@@ -1,20 +1,8 @@
 "use client";
 
 import React, { useEffect } from "react";
-import { useAppSelector, useAppDispatch } from "@/redux/hooks"; // Ensure this path is correct
-import { fetchTop20Coins } from "@/redux/features/marketSlice"; // Ensure this path is correct
 
 function Table() {
-  const dispatch = useAppDispatch();
-  const coins = useAppSelector((state) => state.coins.coins);
-  const status = useAppSelector((state) => state.coins.status);
-  const error = useAppSelector((state) => state.coins.error);
-
-  useEffect(() => {
-    if (status === "idle") {
-      dispatch(fetchTop20Coins());
-    }
-  }, [dispatch, status]);
   return (
     <>
       <table>
